Teme/Tema3/requirements.txt
paho-mqtt==1.5.0
influxdb==5.2.3
influxdb_client